Added axis-aligned bounding boxes, whitted tracer, reflective material (mirror reflection/perfect specular), new shapes (triangle, disk, cylinder, open cylinder), support of compound objects, support for instancing, support for matrix transformations, and other things

+++ Tracer_2/Lambertian.h
@@ -62,7 +62,7 @@ inline void Lambertian::setCD(const float c)
 
 inline RGBColor Lambertian::f(const ShadeRec& sr, const Vector3D& wo, const Vector3D& wi) const
 {
-	return kd * cd * INV_PI;
+	return kd * cd * (float)INV_PI;
 }
